Rename ActivityDetailsContext to the SurveyBasicContext
