+++ src/pages/PublicActivityDetails/index.tsx
@@ -1,8 +1,8 @@
 import { useParams } from 'react-router-dom';
 
-import { Box } from '~/shared/ui';
+import Box from '~/shared/ui/Box';
 import { useCustomTranslation } from '~/shared/utils';
-import { ActivityDetailsContext, ActivityDetailsWidget } from '~/widgets/ActivityDetails';
+import { SurveyBasicContext, ActivityDetailsWidget } from '~/widgets/ActivityDetails';
 
 function PublicActivityDetailsPage() {
   const { appletId, activityId, eventId, publicAppletKey } = useParams();
@@ -14,11 +14,11 @@ function PublicActivityDetailsPage() {
 
   return (
     <Box display="flex" flex={1}>
-      <ActivityDetailsContext.Provider
+      <SurveyBasicContext.Provider
         value={{ appletId, activityId, eventId, publicAppletKey, isPublic: true }}
       >
         <ActivityDetailsWidget />
-      </ActivityDetailsContext.Provider>
+      </SurveyBasicContext.Provider>
     </Box>
   );
 }
