+++ src/widgets/ActivityDetails/index.ts
@@ -1,2 +1,2 @@
 export * from './ui';
-export { ActivityDetailsContext } from './lib';
+export { SurveyBasicContext } from './lib';
