+++ src/widgets/ActivityDetails/lib/index.ts
@@ -1,1 +1,1 @@
-export * from './ActivityDetailsContext';
+export * from './SurveyBasicContext';
